requirements.txt
decorator==4.0.6
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-postgrespool==0.3.0
django-tastypie==0.12.2
django-toolbelt==0.0.1
gunicorn==19.3.0
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
pyjade==3.1.0
python-dateutil==2.4.2
python-mimeparse==0.1.4
requests==2.7.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.8
static3==0.6.1
traitlets==4.0.0
wheel==0.24.0
whitenoise==2.0.4
